atualizando showDB e trazendo leaflet-realtime

+++ showDB_2/printDB_2/views.py
@@ -24,8 +24,8 @@ def batata(request):
     lat = [row[2] for row in cursor.fetchall()]
     cursor.execute('SELECT * FROM teste_3 ORDER BY id DESC LIMIT 1')
     lon = [row[3] for row in cursor.fetchall()]
-    lat_i = int(lat[0])
-    lon_i = int(lon[0])
+    lat_i = float(lat[0])
+    lon_i = float(lon[0])
     db.close()
     context = {'lat_i': lat_i, 'lon_i': lon_i}
 #    return HttpResponse(lat_i)
